fix: template variables bug
I fix a problem within the views that wasn't allowing me to show the correct info in the template. I was creating a list that had the information of the book already displayed but i fixed it deleting the list and only working with the queryset created i could access to the info in the template in a correct way. 

Other fixes are style changes related to the name of url, since I have to access only a book, i decided to change the url from books to book.