feat: bookr logo, bookr reviews logo and styles.
I started by modifying the project's base.html. Then i created a reviews' base.html and programmed it so that when we access book reviews section, the logo changes to bookr reviews.

I added the static files to review so it can serve the logo Bookr reviews to the html. Then i changed the  book_details.html and book_list.html in the namespaced reviews template so it can refer to the base.html, i also moved the base.html and search_result.html to the namespaced folder.

Then i created a static directory for the project, moved the styles from the project's html to main.css so it can be more organized. Also added a new Bookr logo in the static folder. Finally, I modified the project's base.html to show the Bookr logo.